bot/requirements_bot.txt
aiofiles==23.2.1
aiogram==3.5.0
aiohttp==3.9.5
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
async-timeout==4.0.3
attrs==25.3.0
Babel==2.9.1
certifi==2025.4.26
charset-normalizer==3.4.1
colorama==0.4.6
ffmpeg-python==0.2.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
future==1.0.0
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
magic-filter==1.0.12
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
openai-whisper==20240930
propcache==0.3.1
pydantic==2.7.4
pydantic_core==2.18.4
pytz==2025.2
regex==2024.11.6
requests==2.32.3
sympy==1.13.3
tiktoken==0.9.0
torch==2.7.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
